refactor: update appolo clients types
