+++ packages/shop-mobile-expo/src/client/types.ts
@@ -1,7 +1,17 @@
-//  This file will contain common types & interfaces that arguments will use
+// This file will contain common types & interfaces that arguments will use
 import type { MaybeType, ScalarsInterface } from '../types/index';
 
-export interface GeoLocationInterface {
+export interface LocationInterface {
+	type: ScalarsInterface['String'];
+	coordinates: ScalarsInterface['Float'][];
+}
+
+export interface GeolocationCoordinatesInterface {
+	lng: ScalarsInterface['Float'];
+	lat: ScalarsInterface['Float'];
+}
+
+export interface GeolocationInterface {
 	_id?: MaybeType<ScalarsInterface['String']>;
 	id?: MaybeType<ScalarsInterface['String']>;
 	_createdAt?: MaybeType<ScalarsInterface['Date']>;
@@ -15,48 +25,22 @@ export interface GeoLocationInterface {
 	house: ScalarsInterface['String'];
 	postcode?: MaybeType<ScalarsInterface['String']>;
 	notes?: MaybeType<ScalarsInterface['String']>;
-	loc: LocInterface;
-	coordinates?: GeoLocationCoordinatesInterface;
-}
-
-export interface GeoLocationCoordinatesInterface {
-	lng: ScalarsInterface['Float'];
-	lat: ScalarsInterface['Float'];
-}
-
-export interface LocInterface {
-	type: ScalarsInterface['String'];
-	coordinates: ScalarsInterface['Float'][];
-}
-
-export interface GeoLocationCreateInputInterface {
-	countryId: ScalarsInterface['Int'];
-	city: ScalarsInterface['String'];
-	streetAddress: ScalarsInterface['String'];
-	house: ScalarsInterface['String'];
-	postcode?: MaybeType<ScalarsInterface['String']>;
-	notes: MaybeType<ScalarsInterface['String']>;
-	loc: LocInputInterface;
-}
-
-export interface LocInputInterface {
-	type: ScalarsInterface['String'];
-	coordinates: ScalarsInterface['Float'][];
-}
-
-export interface GeoLocationInputInterface {
-	countryId?: MaybeType<ScalarsInterface['Int']>;
-	countryName?: MaybeType<ScalarsInterface['String']>;
-	city?: MaybeType<ScalarsInterface['String']>;
-	streetAddress?: MaybeType<ScalarsInterface['String']>;
-	house?: MaybeType<ScalarsInterface['String']>;
-	postcode?: MaybeType<ScalarsInterface['String']>;
 	loc: LocationInterface;
+	coordinates?: GeolocationCoordinatesInterface;
 }
 
-export interface LocationInterface {
-	type: ScalarsInterface['String'];
-	coordinates: ScalarsInterface['Float'][];
+export interface GeolocationInputInterface {
+	city?: ScalarsInterface['String'];
+	countryId?: ScalarsInterface['Int'];
+	house?: ScalarsInterface['String'];
+	loc?: LocationInterface;
+	streetAddress?: ScalarsInterface['String'];
+	postcode?: MaybeType<ScalarsInterface['String']>;
+}
+
+export interface PagingSortInputInterface {
+	field: ScalarsInterface['String'];
+	sortBy: ScalarsInterface['String'];
 }
 
 export interface PagingOptionsInputInterface {
@@ -65,9 +49,9 @@ export interface PagingOptionsInputInterface {
 	skip?: MaybeType<ScalarsInterface['Int']>;
 }
 
-export interface PagingSortInputInterface {
-	field: ScalarsInterface['String'];
-	sortBy: ScalarsInterface['String'];
+export interface TranslateInterface {
+	locale: ScalarsInterface['String'];
+	value: ScalarsInterface['String'];
 }
 
 export interface ImageInterface {
@@ -78,7 +62,107 @@ export interface ImageInterface {
 	orientation: ScalarsInterface['Int'];
 }
 
-export interface TranslateInterface {
-	locale: ScalarsInterface['String'];
-	value: ScalarsInterface['String'];
+export interface DeviceInterface {
+	_id: ScalarsInterface['String'];
+	id: ScalarsInterface['String'];
+	channelId?: MaybeType<ScalarsInterface['String']>;
+	type: ScalarsInterface['String'];
+	uuid: ScalarsInterface['String'];
+	language?: MaybeType<ScalarsInterface['String']>;
+}
+
+// ENTITIES
+export interface UserInterface {
+	_id: ScalarsInterface['String'];
+	id: ScalarsInterface['String'];
+	geoLocation: GeolocationInterface;
+	apartment: ScalarsInterface['String'];
+	firstName?: MaybeType<ScalarsInterface['String']>;
+	lastName?: MaybeType<ScalarsInterface['String']>;
+	email?: MaybeType<ScalarsInterface['String']>;
+	phone?: MaybeType<ScalarsInterface['String']>;
+	devicesIds: ScalarsInterface['String'][];
+	devices: DeviceInterface[];
+	image?: MaybeType<ScalarsInterface['String']>;
+	fullAddress?: MaybeType<ScalarsInterface['String']>;
+	createdAt?: MaybeType<ScalarsInterface['Date']>;
+	isBanned: ScalarsInterface['Boolean'];
+}
+
+export interface UserLoginInfoInterface {
+	user: UserInterface;
+	token: ScalarsInterface['String'];
+}
+
+/**
+ * Invite structure
+ */
+export interface InviteInterface {
+	_id: ScalarsInterface['String'];
+	id: ScalarsInterface['String'];
+	code: ScalarsInterface['String'];
+	apartment: ScalarsInterface['String'];
+	geoLocation: GeolocationInterface;
+}
+
+export interface NewInviteInterface {
+	__typename: string;
+	apartment: string | number;
+	code: string | number;
+	geoLocation: {
+		__typename: string;
+		city: string;
+		coordinates: {
+			__typename: string;
+			lat: number;
+			lng: number;
+		};
+		countryId: number;
+		countryName: string;
+		createdAt: string;
+		house: string;
+		id: string;
+		notes: string | null;
+		postcode: string | null;
+		streetAddress: string;
+		updatedAt: string;
+	};
+	id: string;
+}
+
+export interface ProductInterface {
+	_id: ScalarsInterface['String'];
+	id: ScalarsInterface['String'];
+	title: TranslateInterface[];
+	description: TranslateInterface[];
+	details: TranslateInterface[];
+	images: ImageInterface[];
+	categories?: MaybeType<MaybeType<ScalarsInterface['String']>[]>;
+	detailsHTML: TranslateInterface[];
+	descriptionHTML: TranslateInterface[];
+	_createdAt?: MaybeType<ScalarsInterface['Date']>;
+	_updatedAt?: MaybeType<ScalarsInterface['Date']>;
+}
+
+export interface WarehouseProductInterface {
+	id: ScalarsInterface['String'];
+	_id: ScalarsInterface['String'];
+	price: ScalarsInterface['Int'];
+	initialPrice: ScalarsInterface['Int'];
+	count?: MaybeType<ScalarsInterface['Int']>;
+	soldCount: ScalarsInterface['Int'];
+	product: ProductInterface;
+	isManufacturing?: MaybeType<ScalarsInterface['Boolean']>;
+	isCarrierRequired?: MaybeType<ScalarsInterface['Boolean']>;
+	isDeliveryRequired?: MaybeType<ScalarsInterface['Boolean']>;
+	isTakeaway?: MaybeType<ScalarsInterface['Boolean']>;
+	deliveryTimeMin?: MaybeType<ScalarsInterface['Int']>;
+	deliveryTimeMax?: MaybeType<ScalarsInterface['Int']>;
+}
+
+export interface ProductInfoInterface {
+	warehouseProduct: WarehouseProductInterface;
+	distance: ScalarsInterface['Float'];
+	warehouseId: ScalarsInterface['String'];
+	warehouseLogo: ScalarsInterface['String'];
 }
